+++ src/helpers/toPath.ts
@@ -18,7 +18,7 @@ const toPathArray = (input: string) => {
     if (part.includes('[')) {
       const { 0: firstPart, 1: temp } = part.split('[');
       result.push(firstPart);
-      const { 0: secondPart } = temp.split(']')[0];
+      const { 0: secondPart } = temp.split(']');
       result.push(secondPart);
     } else { result.push(part); }
   });